requirements-light.txt
streamlit==1.52.2
boto3==1.40.61
botocore==1.40.61
python-dotenv==1.2.1
geopy==2.4.1
